Sketch encode for sequences
This mainly runs into a main limitation on imitation where the
`ImageEncoder` necessarily consumes the encoder. Also we have many
different ways of handling profiles. For TIFF it is quite clear that the
surface for encoding profiles-per-image should exposed. For avif this
needs to be clarified with the underlying format. For gif it seems that
imagemagick does .. magic .. to store different frames under different
profiles in a way that not even Chrome/Firefox support. For webp we find
that only the last image's profile is stored.